// src/connections/item/operations/item/connectionOperationItemRequestBuilderPatchRequestConfiguration.ts
import {RequestOption} from '@microsoft/kiota-abstractions';

export interface ConnectionOperationItemRequestBuilderPatchRequestConfiguration {
    /**
     * Request headers
     */
    headers?: Record<string, string[]> | undefined;
    /**
     * Request options
     */
    options?: RequestOption[] | undefined;
}